refac: bin_tree mod
