+++ crates/tlb-ton/src/lib.rs
@@ -1,9 +1,8 @@
-mod address;
-pub mod bin_tree_aug;
 pub mod bin_tree;
-mod boc;
 pub mod currency;
 pub mod hashmap;
+mod address;
+mod boc;
 mod message;
 mod state_init;
 mod timestamp;
